Add ADIOS conversion job
Adding a separate job to convert ADIOS files, if any, to NetCDF.
The conversion job is added as a dependent job to the model job
and runs once the model run completes.

The conversion job modifies the model mpirun command to run
the ADIOS convertion tool. The conversion job is added only for
E3SM since we currently only need post processing for ADIOS
iotype (not supported by CESM).